rename: Pagination cursor 관련 파일 디렉터리 이동
